backend-api/requirements-docker.txt
# ============================================
# Core Deep Learning
# ============================================
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# ============================================
# Data & Science
# ============================================
numpy==2.3.3
pandas==2.3.3
pillow==11.3.0
scikit-learn==1.7.2
scipy==1.16.3

# ============================================
# Visualization
# ============================================
matplotlib==3.10.7
seaborn==0.13.2

# ============================================
# API
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# Config & Utils
# ============================================
pyyaml==6.0.3
tqdm==4.67.1

# ============================================
# HTTP Requests (pour healthcheck)
# ============================================
requests==2.32.3